customized/python_server/requirements.txt
attrs==21.2.0
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
Pillow==8.3.1
pyrsistent==0.18.0
python-dotenv==0.15.0
six==1.16.0
uuid==1.30
Werkzeug==1.0.1
